requirements.txt
aiobotocore==2.16.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiokafka==0.12.0
aiosignal==1.3.2
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
asgi-lifespan==2.1.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
billiard==4.2.1
botocore==1.35.88
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
elastic-transport==8.15.1
elasticsearch==8.17.0
fastapi==0.115.6
flake8==7.1.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
jmespath==1.0.1
kombu==5.4.2
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.1
pycodestyle==2.12.1
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.25.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.0
yarl==1.18.3
